Webengine improvements [DISCL-388]
- Plugins enabled (flash)
- Replace "select" system drop-down lists with html equivalents (as was
  already done for the Webkit webbrowser)
- Transmit full navigation history via Deflect

+++ tide/core/scene/WebbrowserHistory.h
@@ -56,7 +56,7 @@ public:
 #ifdef TIDE_USE_QT5WEBKITWIDGETS
     explicit WebbrowserHistory( const QWebHistory& history );
 #endif
-    WebbrowserHistory( std::vector<QString>&& items, int currentItemIndex );
+    WebbrowserHistory( std::vector<QString>&& items, size_t currentItemIndex );
 
     const std::vector<QString>& items() const;
     size_t currentItemIndex() const;
